requirements.txt
## The following requirements were added by pip freeze:
Django==2.2.6
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.3
Markdown==3.1.1
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
